# Improve lexing performance by generating lexer with Ragel 7.0.0.9 and avoiding lambda in token emission