While investigating why query lexing was measurably slower in JRuby than with MRI (see https://github.com/jruby/jruby/issues/4623 ) I noticed that GraphQL's lexer seemed to be generated with Ragel 6.

I then noticed that this version was rather old and that a newer Ragel 7 seemed to have a lot of improvements on the Ruby output backend, even adding four different output options for generated lexers:

* `-T0` Binary search (default)
* `-T1` Binary search with expanded actions
* `-F0` Flat table
* `-F1` Flat table with expanded actions

Testing all four of these combinations and comparing with the current GraphQL lexer on both MRI and JRuby showed that for a lexing-only benchmark (included below), Ragel 7's `-F1` improved MRI peformance by 87% and JRuby performance by 263%:

* MRI before:

```
Running with ruby 2.4.1p111 (2017-03-22 revision 58053) [x86_64-linux]
Warming up --------------------------------------
               Lexer   111.000  i/100ms
Calculating -------------------------------------
               Lexer      1.124k (± 3.9%) i/s -     33.744k in
30.069500s
```

* MRI after:

```
Running with ruby 2.4.1p111 (2017-03-22 revision 58053) [x86_64-linux]
Warming up --------------------------------------
               Lexer   209.000  i/100ms
Calculating -------------------------------------
               Lexer      2.102k (± 3.0%) i/s -     63.118k in
30.052143s
```

* JRuby before:

```
Running with jruby 9.1.9.0 (2.3.3) 2017-05-15 28aa830 Java HotSpot(TM)
64-Bit Server VM 25.131-b11 on 1.8.0_131-b11 +jit [linux-x86_64]
Warming up --------------------------------------
               Lexer    24.000  i/100ms
Calculating -------------------------------------
               Lexer    250.435  (± 4.8%) i/s -      7.512k in
30.080584s
```

* JRuby after:

```
Running with jruby 9.1.9.0 (2.3.3) 2017-05-15 28aa830 Java HotSpot(TM)
64-Bit Server VM 25.131-b11 on 1.8.0_131-b11 +jit [linux-x86_64]
Warming up --------------------------------------
               Lexer    88.000  i/100ms
Calculating -------------------------------------
               Lexer    910.242  (± 3.0%) i/s -     27.280k in
29.999337s
```

Unfortunately, this does not yet bring JRuby performance to MRI's level, and I plan to continue looking at how to improve this.

An important note is that the latest Ragel 7 release drops support for a lot of languages, unfortunately including Ruby (see https://www.colm.net/news/2016/07/31/ragel-target-langs.html for the announcement) so the last currently-available Ragel 7 version that can be used is 7.0.0.9, which was the version I used.

Lexing benchmark used to obtain the numbers above:

```ruby
require 'graphql'
require 'benchmark/ips'

input_string = "{
  viewer {
    login
    starredRepositories {
      totalCount
    }
    repositories(first: 3) {
      edges {
        node {
          name
          stargazers {
            totalCount
          }
          forks {
            totalCount
          }
          watchers {
            totalCount
          }
          issues(states:[OPEN]) {
            totalCount
          }
        }
      }
    }
  }
}
"

puts "Running with #{RUBY_DESCRIPTION}"

Benchmark.ips do |benchmark|
  benchmark.time = 30
  benchmark.warmup = 30
  benchmark.report('Lexer') {
GraphQL::Language::Lexer.tokenize(input_string) }
  benchmark.compare!
end
```

# Update: Also get rid of lambda for emitting tokens

This has a small impact on MRI:

```
Running with ruby 2.4.1p111 (2017-03-22 revision 58053) [x86_64-linux]
Warming up --------------------------------------
               Lexer   211.000  i/100ms
Calculating -------------------------------------
               Lexer      2.121k (± 3.3%) i/s -     63.722k in  30.080212s
```

But a **HUGE** impact on JRuby:

```
Running with jruby 9.1.9.0 (2.3.3) 2017-05-15 28aa830 Java HotSpot(TM)
64-Bit Server VM 25.131-b11 on 1.8.0_131-b11 +jit [linux-x86_64]
Warming up --------------------------------------
               Lexer   372.000  i/100ms
Calculating -------------------------------------
               Lexer      4.078k (± 5.0%) i/s -    122.016k in  30.008956s
```

Especially when using the `-Xcompile.invokedynamic=true` option:

```
Running with jruby 9.1.9.0 (2.3.3) 2017-05-15 28aa830 Java HotSpot(TM)
64-Bit Server VM 25.131-b11 on 1.8.0_131-b11 +indy +jit [linux-x86_64]
Warming up --------------------------------------
               Lexer   800.000  i/100ms
Calculating -------------------------------------
               Lexer      9.649k (± 7.9%) i/s -    287.200k in  30.009197s
```

Note that in the master branch, JRuby with `-Xcompile.invokedynamic=true` is the worst performing option (clocking in at about 141k i/s), so with this PR it goes from slowest option to being the fastest by far, with a speedup of about 68x.