@rmosolgo 2017-05-26
Wow, thank you so much for looking into this! I saw the news about removing Ruby support from Ragel 7, so I thought we would just stay on Ragel 6 forever. Now, I see that the earliest versions of Ragel 7 _do_ have Ruby support, so we can at least use _that_ for a while! 

Also, reading up on your exploring with JRuby, let me try to refactor out that lambda. I think it can be handled a bit better, I really didn't understand Ragel's setup well when I wrote this the first time!